+++ to-esm.js
@@ -1,0 +1,10 @@
+const {transform} = require("cjstoesm");
+
+(async () => {
+    //console.log(__dirname)
+    await transform({
+        cwd: __dirname + '//src',
+        input: "**/*.*",
+        outDir: "../dist/esm"
+    });
+})()